+++ R/counts_statistics.R
@@ -16,27 +16,21 @@
 #' @export
 plot_gene_detection <- function(dds, thresholds = c(3, 10, 20, 50)) {
 
-  # prevent 'no visible binding for global variable' package warnings
-  gene_id <- count <- sample_id <- detection_threshold <- number_of_genes <- fraction_of_genes <- fraction_of_reads <- NULL
+  # prevent integer overflow when adding very large counts
+  counts <- assay(dds)
+  mode(counts) <- "double"
 
-  assay(dds) %>%
-    as_tibble(rownames = "gene_id") %>%
-    pivot_longer(-gene_id, names_to = "sample_id", values_to = "count") %>%
-    arrange(-count) %>%
-    group_by(sample_id) %>%
-    summarize(
-      detection_threshold = count,
-      number_of_genes = row_number(),
-      .groups = "drop"
-    ) %>%
-    group_by(sample_id, detection_threshold) %>%
-    summarize(number_of_genes = max(number_of_genes), .groups = "drop") %>%
-    filter(detection_threshold %in% thresholds) %>%
-    mutate(detection_threshold = factor(detection_threshold)) %>%
+  thresholds %>%
+    map_dfr(function(threshold){
+      colSums(counts >= threshold) %>%
+        enframe(name = "sample_id", value = "number_of_genes") %>%
+        mutate(detection_threshold = threshold)
+    }) %>%
+    mutate(detection_threshold = factor(detection_threshold, levels = sort(thresholds))) %>%
     ggplot(aes(detection_threshold, number_of_genes, group = sample_id, label = sample_id)) +
-    geom_path(alpha = .4) +
-    labs(x = "detection threshold", y = "number of genes") +
-    cowplot::theme_cowplot()
+      geom_path(alpha = 0.4) +
+      labs(x = "detection threshold", y = "number of genes") +
+      cowplot::theme_cowplot()
 }
 
 #' Plot total counts per sample
@@ -56,20 +50,31 @@ plot_gene_detection <- function(dds, thresholds = c(3, 10, 20, 50)) {
 #' @export
 plot_total_counts <- function(dds, n_bins = 50) {
 
-  # prevent 'no visible binding for global variable' package warnings
-  x <- NULL
+  # prevent integer overflow when adding very large counts
+  counts <- assay(dds)
+  mode(counts) <- "double"
 
-  data.frame(x = colSums(assay(dds))) %>%
-    ggplot(aes(x = x)) +
+  data.frame(col_sum = colSums(counts)) %>%
+    ggplot(aes(x = col_sum)) +
     geom_histogram(bins = n_bins) +
     labs(x = "total count", y = "number of samples") +
     cowplot::theme_cowplot()
+}
+
+# returns for each element in y the position of the
+# largest element in x that is less or equal to y
+pos_largest_lower_bound <- function(x, y){
+  map_dbl(y, function(y){
+    res <- which.max(x[x <= y])
+    ifelse(length(res)==0,NA,res)
+  })
 }
 
 #' Plot the library complexity
 #'
 #' Plot per sample the fraction of genes, versus the fraction of total counts.
 #' @param dds A DESeqDataSet
+#' @param show_progress Whether to show a progress bar of the computation.
 #'
 #' @return A ggplot object of the ggplot2 package that contains the library complexity plot.
 #'
@@ -80,25 +85,34 @@ plot_total_counts <- function(dds, n_bins = 50) {
 #' plot_library_complexity(dds)
 #'
 #' @export
-plot_library_complexity <- function(dds) {
+plot_library_complexity <- function(dds, show_progress = TRUE) {
 
-  # prevent 'no visible binding for global variable' package warnings
-  gene_id <- count <- sample_id <- fraction_of_genes <- fraction_of_counts <- NULL
+  # prevent integer overflow when adding very large counts
+  counts <- assay(dds)
+  mode(counts) <- "double"
 
-  purrr::map_dfr(1:ncol(dds), function(i) {
-    cts <- sort(assay(dds)[,i], decreasing=T)
+  df <- map_dfr(1:ncol(dds), function(i) {
+    cts <- sort(counts[, i], decreasing = T)
+    fraction_of_genes <- seq_along(cts) / length(cts)
+    fraction_of_counts <- cumsum(cts) / sum(cts)
+    gene_frac_subset_pos <- c(which(fraction_of_genes < 0.01), pos_largest_lower_bound(fraction_of_genes, seq(.01, 1, .01)))
+
     tibble(
       sample_id = SummarizedExperiment::colnames(dds)[i],
-      fraction_of_counts = cumsum(cts)/sum(cts),
-      fraction_of_genes = (1:length(cts))/length(cts)
+      fraction_of_genes = fraction_of_genes[gene_frac_subset_pos],
+      fraction_of_counts = fraction_of_counts[gene_frac_subset_pos]
     )
-  }) %>%
+  }, .progress = show_progress) %>%
+    filter(stats::complete.cases(.))
+
+
+  df %>%
     ggplot(aes(fraction_of_genes, fraction_of_counts, group = sample_id, label = sample_id)) +
-      geom_line(alpha = .2) +
-      theme_minimal() +
-      scale_x_sqrt(breaks = c(.01, .05, .1, .2, .5, 1)) +
-      scale_y_continuous(breaks = seq(0, 1, .2)) +
-      labs(x = "fraction of genes", y = "fraction of counts") +
-      theme(legend.position = "none") +
-      cowplot::theme_cowplot()
+    geom_line(alpha = 0.2) +
+    theme_minimal() +
+    scale_x_sqrt(breaks = c(0.01, 0.05, 0.1, 0.2, 0.5, 1)) +
+    scale_y_continuous(breaks = seq(0, 1, 0.2)) +
+    labs(x = "fraction of genes", y = "fraction of counts") +
+    theme(legend.position = "none") +
+    cowplot::theme_cowplot()
 }