+++ NEWS.md
@@ -1,3 +1,13 @@
+# RNAseqQC 0.2
+* Add the possibility to plot gene expression in plot_pca()
+* plot_pca() and plot_pca_scatters() now detect if the variable to color by is
+continuous or discrete and choose an appropriate, colorblind friendly color scale
+* Add option to plot rasterised points in plot_pca_scatters()
+* It is now possible to plot squared loadings in the plot_loadings() function
+* Fix a bug where the distance parameter was ignored in plot_sample_clustering()
+* Fix a bug where the annotate_top_n parameter didn't work in plot_loadings() when
+the input was a matrix
+
 # RNAseqQC 0.1.4
 * Fix of bug in plot_pca() that was introduced in the previous version, causing
 the the feature selection to be circumvented.